Move base health service to internal
