+++ pkg/suggestion/v1alpha3/chocolate/chocolate_service.py
@@ -8,7 +8,7 @@ from pkg.suggestion.v1alpha3.internal.constant import DOUBLE
 from pkg.suggestion.v1alpha3.internal.search_space import HyperParameterSearchSpace
 from pkg.suggestion.v1alpha3.internal.trial import Trial, Assignment
 from pkg.suggestion.v1alpha3.chocolate.base_chocolate_service import BaseChocolateService
-from pkg.suggestion.v1alpha3.base_health_service import HealthServicer
+from pkg.suggestion.v1alpha3.internal.base_health_service import HealthServicer
 
 logger = logging.getLogger(__name__)
