+++ pkg/suggestion/v1alpha3/skopt/base_skopt_service.py
@@ -2,7 +2,7 @@ import datetime
 import logging
 import skopt
 
-from pkg.suggestion.v1alpha3.internal.constant import *
+from pkg.suggestion.v1alpha3.internal.constant import INTEGER, DOUBLE, CATEGORICAL, DISCRETE, MAX_GOAL
 from pkg.suggestion.v1alpha3.internal.trial import Assignment
 
 logger = logging.getLogger(__name__)
