+++ test/suggestion/v1alpha3/test_chocolate_service.py.failed
@@ -5,7 +5,7 @@ import unittest
 from pkg.apis.manager.v1alpha3.python import api_pb2_grpc
 from pkg.apis.manager.v1alpha3.python import api_pb2
 
-from pkg.suggestion.v1alpha3.chocolate_service import ChocolateService
+from pkg.suggestion.v1alpha3.chocolate.chocolate_service import ChocolateService
 
 
 class TestHyperopt(unittest.TestCase):
